restructure the package for working in a catkin workspace
