+++ .gitignore
@@ -1,6 +1,3 @@
-checkpoint/
-pretrained/
-
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
@@ -103,3 +100,6 @@ ENV/
 
 # mypy
 .mypy_cache/
+
+checkpoint/
+pretrained/
