+++ CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 2.8.3)
+project(pytorch_retinanet)
+
+find_package(catkin REQUIRED)
+catkin_package()
+catkin_python_setup()
