--- run_test.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-cd scripts
-
-python test.py \
-    --img_path "./data/voc_data/voc_test/000001.jpg"
-
-cd -
-
