--- run_train.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-cd scripts
-
-python3 train.py
-
-cd -
