+++ setup.py
@@ -1,0 +1,10 @@
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+d = generate_distutils_setup(
+    packages=[
+        'pytorch_retinanet'
+    ],
+    package_dir={'': 'src'},
+)
+setup(**d)
